transfer uniform_random_batch_size_like api
