+++ python/paddle/fluid/layers/distributions.py
@@ -221,8 +221,10 @@ class Uniform(Distribution):
             zero_tmp = tensor.fill_constant_batch_size_like(
                 self.low + self.high, batch_shape + shape, self.low.dtype, 0.0
             )
-            uniform_random_tmp = nn.uniform_random_batch_size_like(
-                zero_tmp, zero_tmp.shape, min=0.0, max=1.0, seed=seed
+            uniform_random_tmp = (
+                paddle.tensor.random.uniform_random_batch_size_like(
+                    zero_tmp, zero_tmp.shape, min=0.0, max=1.0, seed=seed
+                )
             )
             output = (
                 uniform_random_tmp * (zero_tmp + self.high - self.low)
