+++ python/paddle/fluid/tests/unittests/test_uniform_random_bf16_op.py
@@ -23,6 +23,7 @@ from paddle.fluid.tests.unittests.test_uniform_random_op import (
     output_hist,
     output_hist_diag,
 )
+from paddle.tensor import random
 
 
 class TestUniformRandomOpBF16(OpTest):
@@ -262,7 +263,7 @@ class TestUniformRandomBatchSizeLikeOpBF16API(unittest.TestCase):
         train_program = fluid.Program()
         with fluid.program_guard(train_program, startup_program):
             input = fluid.data(name="input", shape=[1, 3], dtype='uint16')
-            out_1 = fluid.layers.uniform_random_batch_size_like(
+            out_1 = random.uniform_random_batch_size_like(
                 input, [2, 4], dtype=np.uint16
             )  # out_1.shape=[1, 4]
